+++ src/CutCode.CrossPlatform/ViewModels/AddViewModel.cs
@@ -7,11 +7,9 @@ using System.Text;
 using System.Threading.Tasks;
 using Avalonia;
 using Avalonia.Media;
-using CutCode.CrossPlatform.Interfaces;
 using CutCode.CrossPlatform.Models;
 using CutCode.CrossPlatform.ViewModels;
 using CutCode.CrossPlatform.Views;
-using CutCode.CrossPlatform.DataBase;
 using Newtonsoft.Json;
 using ReactiveUI;
